provides multiple accessors to word-level features.
