+++ pyproject.toml
@@ -46,6 +46,7 @@ greek-accentuation = "^1.2.0"
 python-Levenshtein = "^0.12.0"
 stanza = "^1.0.0"
 nltk = "^3.5"
+stringcase = "^1.2"
 
 [tool.poetry.dev-dependencies]
 pytest = "^6.1.1"
